requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
requests
django==2.2.6
pytest
pytest-pythonpath==0.7.3
pytest-django
djangorestframework
djangorestframework_simplejwt
